Have a common post or comment component
