+++ src/components/ui/index.js
@@ -2,5 +2,6 @@ import NavLink from "./NavLink";
 import FormErrors from "./FormErrors";
 import HeaderText from "./HeaderText";
 import Loader from "./Loader";
+import PostOrComment from "./PostOrComment";
 
-export { NavLink, FormErrors, HeaderText, Loader };
+export { NavLink, FormErrors, HeaderText, Loader, PostOrComment };
